Move dataclasses to separate module
