+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "khinsider-downloader"
-version = "0.1.7"
+version = "0.2.0-dev1"
 description = ""
 authors = [{ name = "NovaHFly", email = "mugiviska@gmail.com" }]
 requires-python = "~=3.12"
@@ -12,12 +12,6 @@ dependencies = [
     "lxml>=5.3.0,<6",
     "tenacity>=9.0.0,<10",
 ]
-
-[tool.hatch.build.targets.sdist]
-include = ["khinsider"]
-
-[tool.hatch.build.targets.wheel]
-include = ["khinsider"]
 
 [build-system]
 requires = ["hatchling"]
